make the server just work part two
Sorry for the last two commit messages
I just wanted to get the server up to a usable state.

Before this PR, the server couldn't be initialized successfuly.

Luckily, some good soul fixed the issue,
https://github.com/tailwindlabs/tailwindcss-intellisense/pull/263

but the maintainer didn't published yet a release with that fix,
so I manually compiled the tailwindcss-intellisense repo,
and included the fix with this PR.

So everything should be good now.