fully automate the download extension script
the script will now use the extension package.json file
and automatically generate the package-lock.json file

So now anyone can update the server with the script.

Before this change. When someone run the script,
the script would unintentionally delete language-server/package.json and
language-server/package-lock.json files. So I would have to discard the deletion
of the two files, and change the version in the package.json

Now. nothing of that is required. Just run the script. and it will do the rest.

Hopefully the tailwind server will be published on npm so the script will be unnecessary.