Update server and scripts (#10)
* fully automate the download extension script

the script will now use the extension package.json file
and automatically generate the package-lock.json file

So now anyone can update the server with the script.

Before this change. When someone run the script,
the script would unintentionally delete language-server/package.json and
language-server/package-lock.json files. So I would have to discard the deletion
of the two files, and change the version in the package.json

Now. nothing of that is required. Just run the script. and it will do the rest.

Hopefully the tailwind server will be published on npm so the script will be unnecessary.

* update the server

* remove userLanguages as it is no longer necessary

* use only the server files that are needed and fix typos

Co-authored-by: Rafał Chłodnicki <rchl2k@gmail.com>

* add missing setting

* Revert "add missing setting"

This reverts commit 4e6e2a4af2565d0e5c1e7f28fa6b109e5c5f698b.

* one-liner

Co-authored-by: Predrag Nikolic <mail@predrag.pro>
Co-authored-by: Rafał Chłodnicki <rchl2k@gmail.com>