remove is_allowed_to_start
Let the server decide if it can run in the workspace or not.

This way the client extension will not prevent the server from running
in a workspace supported by the server.

Plus the server now bundles the tailwindcss dependency,
That was one of the reasons why we had the check in the first place.
(because without that dependency the server didn't start)